+++ src/main/java/com/angkorteam/fintech/widget/group/GroupPreviewGeneralPanel.java
@@ -82,7 +82,7 @@ public class GroupPreviewGeneralPanel extends Panel {
 
     protected List<ActionItem> savingAccountAction(String s, Map<String, Object> model) {
         List<ActionItem> actions = Lists.newArrayList();
-        Integer status = (Integer) model.get("status");
+        Long status = (Long) model.get("status");
         if (100 == status) {
             actions.add(new ActionItem("Approve", Model.of("Approve"), ItemCss.PRIMARY));
         } else if (200 == status) {
